requirements.txt
gradio>=4.30.0
langchain-core>=0.1.44
python-dotenv>=1.0.1
PyPDF2>=3.0.1
python-docx>=1.1.0
pytesseract>=0.3.10
Pillow>=10.3.0
requests>=2.31.0


